Started adding support for custom profiles +more
Started adding support for custom profiles and dynamic monitor selection
support.

The idea is, if you use the same computer at a desk, and have a TV just
a few meters (/feet) away, and it's connected to a TV (+ maybe a
surround system). For me personally it's preferable to activate the TV
video  and audio (Computer->TV) output, and if I prefer, deactivate the
unsused monitors, so I don't have to "look" for my cursor, because it's
off screen.